fix(semidefinite-face): keep math tests off product dispatch
